+++ src/test/java/org/jabref/logic/autosaveandbackup/BackupManagerTest.java
@@ -16,17 +16,17 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 public class BackupManagerTest {
 
     @Test
-    public void autosaveFileNameIsCorrectlyGeneratedInAppDataDirectory() {
+    public void backupFileNameIsCorrectlyGeneratedInAppDataDirectory() {
         Path bibPath = Path.of("tmp", "test.bib");
-        Path savPath = BackupManager.getBackupPathForNewBackup(bibPath);
+        Path bakPath = BackupManager.getBackupPathForNewBackup(bibPath);
 
-        assertEquals(FileUtil.getAppDataBackupDir(), savPath.getParent());
-        String start = savPath.getFileName().toString().substring(0, 20); // Timestamp will differ
+        assertEquals(FileUtil.getAppDataBackupDir(), bakPath.getParent());
+        String start = bakPath.getFileName().toString().substring(0, 20); // Timestamp will differ
         assertEquals("27182d3c--test.bib--", start);
     }
 
     @Test
-    public void autosaveFileIsEqualForNonExistingBackup() throws Exception {
+    public void backupFileIsEqualForNonExistingBackup() throws Exception {
         Path originalFile = Path.of(BackupManagerTest.class.getResource("no-autosave.bib").toURI());
         assertFalse(BackupManager.backupFileDiffers(originalFile));
     }
@@ -34,8 +34,8 @@ public class BackupManagerTest {
     @Test
     public void backupFileIsEqual() throws Exception {
         // Prepare test: Create backup file on "right" path
-        Path source = Path.of(BackupManagerTest.class.getResource("no-changes.bib.sav").toURI());
-        Path target = FileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI()), BackupFileType.AUTOSAVE);
+        Path source = Path.of(BackupManagerTest.class.getResource("no-changes.bib.bak").toURI());
+        Path target = FileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI()), BackupFileType.BACKUP);
         Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
 
         Path originalFile = Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI());
@@ -45,8 +45,8 @@ public class BackupManagerTest {
     @Test
     public void backupFileDiffers() throws Exception {
         // Prepare test: Create backup file on "right" path
-        Path source = Path.of(BackupManagerTest.class.getResource("changes.bib.sav").toURI());
-        Path target = FileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("changes.bib").toURI()), BackupFileType.AUTOSAVE);
+        Path source = Path.of(BackupManagerTest.class.getResource("changes.bib.bak").toURI());
+        Path target = FileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("changes.bib").toURI()), BackupFileType.BACKUP);
         Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
 
         Path originalFile = Path.of(BackupManagerTest.class.getResource("changes.bib").toURI());
@@ -54,20 +54,20 @@ public class BackupManagerTest {
     }
 
     @Test
-    public void correctBackupFileDeterminedForMultipleSavFiles() throws Exception {
+    public void correctBackupFileDeterminedForMultipleBakFiles() throws Exception {
         // Prepare test: Create backup files on "right" path
 
         // most recent file does not have any changes
-        Path source = Path.of(BackupManagerTest.class.getResource("no-changes.bib.sav").toURI());
-        Path target = FileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI()), BackupFileType.AUTOSAVE);
+        Path source = Path.of(BackupManagerTest.class.getResource("no-changes.bib.bak").toURI());
+        Path target = FileUtil.getPathForNewBackupFileAndCreateDirectory(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI()), BackupFileType.BACKUP);
         Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
 
-        // create "older" .sav files containing changes
+        // create "older" .bak files containing changes
         for (int i = 0; i < 10; i++) {
             source = Path.of(BackupManagerTest.class.getResource("changes.bib").toURI());
             Path directory = FileUtil.getAppDataBackupDir();
             String timeSuffix = "2020-02-03--00.00.0" + Integer.toString(i);
-            String fileName = FileUtil.getUniqueFilePrefix(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI())) + "--no-changes.bib--" + timeSuffix + ".sav";
+            String fileName = FileUtil.getUniqueFilePrefix(Path.of(BackupManagerTest.class.getResource("no-changes.bib").toURI())) + "--no-changes.bib--" + timeSuffix + ".bak";
             target = directory.resolve(fileName);
             Files.copy(source, target, StandardCopyOption.REPLACE_EXISTING);
         }